Action sheet working
Also update build, implementing upload functions next
