+++ src/pages/get-image/get-image.ts
@@ -2,8 +2,6 @@ import { processLintResults } from '@ionic/app-scripts/dist/lint/lint-utils';
 import { Component } from '@angular/core';
 import { Http, Headers } from '@angular/http';
 import { NavController, NavParams, LoadingController } from 'ionic-angular';
-import { ReceiptPoster } from './../shared/receipt-post.service';
-import { OcrUploadImageModel } from './../../models/ocruploadimage.model';
 import { NgForm } from '@angular/forms/src/directives';
 
 // Page import
@@ -14,31 +12,33 @@ import { NgForm } from '@angular/forms/src/directives';
   templateUrl: 'get-image.html'
 })
 export class GetImagePage {
-  model = new OcrUploadImageModel('');
+  // model = new OcrUploadImageModel('');
   
-  ocrreply: any;
-  private resultData: any;
+  // ocrreply: any;
+  // private resultData: any;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, private http: Http, private receiptPoster: ReceiptPoster, private loadingController: LoadingController) { 
-  }
+  // receiptImage: any;
 
-  submitForm(form: NgForm) {
-    let loader = this.loadingController.create({
-      content: 'Processing receipt...'
-    });
+  // constructor(public navCtrl: NavController, public navParams: NavParams, private http: Http) { 
+  // }
 
-    loader.present().then(() => {
-        console.log(form.value);
-        this.receiptPoster.postReceiptForm(this.model)
-        .subscribe(
-          data => this.ocrreply = data.somethingReturned,
-          err => console.log('error: ', err),
-          () => console.log('Something returned: ', this.ocrreply),
-        );
-        loader.dismiss();
-    });
+  // submitForm(form: NgForm) {
+  //   let loader = this.loadingController.create({
+  //     content: 'Processing receipt...'
+  //   });
+
+  //   loader.present().then(() => {
+  //       console.log(form.value);
+  //       this.receiptPoster.postReceiptForm(this.model)
+  //       .subscribe(
+  //         data => this.ocrreply = data.somethingReturned,
+  //         err => console.log('error: ', err),
+  //         () => console.log('Something returned: ', this.ocrreply),
+  //       );
+  //       loader.dismiss();
+  //   });
     
-  }
+  // }
 
   // serviceClicked(event, result){
   //   this.navCtrl.push(YelpResultPage, { result: result });
